Reorganize the project and fix widget drawing.
